+++ sdk/boletoHolmes/boletoHolmes.js
@@ -1,4 +1,5 @@
 const rest = require('../utils/rest.js');
+const check = require('../utils/check.js');
 const Resource = require('../utils/resource.js').Resource
 
 
@@ -33,8 +34,8 @@ class BoletoHolmes extends Resource {
         this.tags = tags;
         this.result = result;
         this.status = status;
-        this.created = created;
-        this.updated = updated;
+        this.created = check.datetime(created);
+        this.updated = check.datetime(updated);
     }
 }